refactor: move ui deployment to it's own file
